+++ CHANGELOG.md
@@ -152,7 +152,9 @@
   - `Spiral\Snapshots\SnapshotterInterface` usage replaced with `Spiral\Exceptions\ExceptionReporterInterface` in all classes.
 - **Other Features**
   - [spiral/debug] Added `Spiral\Debug\StateConsumerInterface`.
-
+  - [spiral/monolog-bridge] Added the ability to configure the default channel using the configuration file or 
+    environment variable `MONOLOG_DEFAULT_CHANNEL`.
+  
 ## v2.12.0 - 2022-04-07
 - **Medium Impact Changes**
   - Bootloaders `Spiral\Bootloader\Broadcast\BroadcastBootloader`, `Spiral\Bootloader\Http\WebsocketsBootloader`